move files and pass down isExerciseGroup
