+++ src/components/exercise-generation/ai-exercise-generation-button.tsx
