+++ src/components/exercise-generation/exercise-generation-wizard/execute-ai-prompt.ts
