+++ src/components/exercise-generation/exercise-generation-wizard/exercise-type.tsx
@@ -1,6 +1,3 @@
-import { useRouter } from 'next/router'
-import { useEffect } from 'react'
-
 import { WizardPageProps } from './wizard-page-props'
 import { MenuButton, MenuItem } from '../menu-button'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
@@ -8,6 +5,7 @@ import { replacePlaceholders } from '@/helper/replace-placeholders'
 
 interface ExerciseTypeProps extends WizardPageProps {
   exerciseType: string | null
+  isExerciseGroup: boolean
   setExerciseType: (type: string | null) => void
   numberOfSubtasks: number
   setNumberOfSubtasks: (num: number) => void
@@ -15,6 +13,7 @@ interface ExerciseTypeProps extends WizardPageProps {
 
 export function ExerciseType({
   exerciseType,
+  isExerciseGroup,
   setExerciseType,
   numberOfSubtasks,
   setNumberOfSubtasks,
@@ -23,14 +22,6 @@ export function ExerciseType({
 }: ExerciseTypeProps) {
   const exerciseTypeStrings =
     useLoggedInData()!.strings.ai.exerciseGeneration.exerciseType
-
-  const isExerciseGroup = useRouter()
-    .asPath.toLowerCase()
-    .includes('/exercisegroup/')
-
-  useEffect(() => {
-    if (!isExerciseGroup) setNumberOfSubtasks(0)
-  }, [isExerciseGroup, setNumberOfSubtasks])
 
   return (
     <div className="flex flex-col">
@@ -95,7 +86,7 @@ export function ExerciseType({
               }}
               className="ml-2 rounded-md border border-brand-200 p-2 pl-2 focus:border-brand-200 focus:outline-brand-700"
               placeholder={exerciseTypeStrings.numberOfSubtasksPlaceholder}
-              autoFocus
+              autoFocus={!isSummary}
             />
           </div>
         </>
