+++ src/components/exercise-generation/exercise-generation-wizard/generate-prompt.ts
