+++ src/components/exercise-generation/exercise-generation-wizard/prompt.tsx
