+++ src/components/exercise-generation/exercise-preview-page.tsx
