+++ src/components/exercise-generation/exercise-test-data.tsx
@@ -1,3 +1,4 @@
+// TODO: @CodingDive is this still needed?
 export const exerciseTestData = {
   heading: 'Dreisatz: Wie weit fährt ein Auto?',
   exercises: [
