+++ src/components/exercise-generation/lazy-load-exercise-generation-wrapper-or-null.tsx
@@ -3,8 +3,10 @@ import { useRouter } from 'next/router'
 import { useState, useEffect } from 'react'
 
 import { ExerciseGenerationWrapperProps } from './exercise-generation-wrapper'
-import { extractTopicFromTitle } from '@/components/content/exercises/exercise-generation-wizard/topic'
-import { useAiFeatures } from '@/components/content/exercises/use-ai-features'
+import { extractTopicFromTitle } from '@/components/exercise-generation/exercise-generation-wizard/topic'
+import { useAiFeatures } from '@/components/exercise-generation/use-ai-features'
+import type { AllowedPluginType } from '@/pages/entity/create/[type]/[taxonomyId]'
+import { TemplatePluginType } from '@/serlo-editor-integration/types/template-plugin-type'
 
 const ExerciseGenerationWrapper = dynamic<ExerciseGenerationWrapperProps>(() =>
   import('./exercise-generation-wrapper').then(
@@ -14,12 +16,16 @@ const ExerciseGenerationWrapper = dynamic<ExerciseGenerationWrapperProps>(() =>
 
 interface LazyLoadExerciseGenerationWrapperOrNullProps {
   subject: string
+  entityType: AllowedPluginType
   taxonomyTitle: string
   setEditorState: ExerciseGenerationWrapperProps['setEditorState']
 }
 
+const allowedExerciseTemplates = ['Exercise', 'ExerciseGroup']
+
 export function LazyLoadExerciseGenerationWrapperOrNull({
   subject,
+  entityType,
   taxonomyTitle,
   setEditorState,
 }: LazyLoadExerciseGenerationWrapperOrNullProps) {
@@ -29,18 +35,18 @@ export function LazyLoadExerciseGenerationWrapperOrNull({
   const topic = extractTopicFromTitle(taxonomyTitle)
 
   useEffect(() => {
-    if (router.query.showAiWizard === 'true' && canUseAiFeatures) {
-      setShowWizard(true)
-    } else {
-      setShowWizard(false)
-    }
-  }, [showWizard, canUseAiFeatures, router.query])
+    const shouldShow =
+      allowedExerciseTemplates.includes(entityType as TemplatePluginType) &&
+      router.query.showAiWizard !== undefined &&
+      canUseAiFeatures
+    setShowWizard(shouldShow)
+  }, [showWizard, canUseAiFeatures, router.query, entityType])
 
-  if (!showWizard) {
-    return null
-  }
+  if (!showWizard) return null
+
   return (
     <ExerciseGenerationWrapper
+      isExerciseGroup={entityType === 'ExerciseGroup'}
       closeWizard={() => {
         setShowWizard(false)
