+++ src/components/exercise-generation/use-ai-features.tsx
