+++ src/serlo-editor/plugins/serlo-template-plugins/exercise-group/text-exercise-group.tsx
@@ -10,8 +10,8 @@ import {
 import { ContentLoaders } from '../helpers/content-loaders/content-loaders'
 import { TextExerciseTypePluginState } from '../text-exercise'
 import { ToolbarMain } from '../toolbar-main/toolbar-main'
-import { AiExerciseGenerationButton } from '@/components/content/exercises/ai-exercise-generation-button'
-import { useAiFeatures } from '@/components/content/exercises/use-ai-features'
+import { AiExerciseGenerationButton } from '@/components/exercise-generation/ai-exercise-generation-button'
+import { useAiFeatures } from '@/components/exercise-generation/use-ai-features'
 import { FaIcon } from '@/components/fa-icon'
 import { useEditorStrings } from '@/contexts/logged-in-data-context'
 import { UuidType } from '@/data-types'
